+++ Mods/lx_ugmtools_b331010b-6678-4373-8cc4-8f004ab0bf36/Story/RawFiles/Lua/Server/_InitServer.lua
@@ -10,6 +10,8 @@ Ext.Require("Server/VisualResources.lua")
 Ext.Require("Server/VisualSetRandomizer.lua")
 Ext.Require("Server/LoseControlFix.lua")
 Ext.Require("Server/Fade.lua")
+Ext.Require("Server/AlignmentFixer.lua")
+Ext.Require("Server/ModListWatchdog.lua")
 -- Ext.Require("Server/ClickTools.lua")