+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 GM Tools : Ultimate GM Toolbox Changelog
 
+# 0.11.11.20
+New
+* Included a host mod list modification warning. That way, people that don't know about the lack of sync between client and server mod list introduced by Gift Bag 2 won't be screwed by it
+* Included a base GM faction fixer, since something is sometimes causing relations between factions to change
+
 # 0.11.10.19
 New
 * Included an experimental fix for LoseControl statuses when the GM is possessing the characters